+++ SPL2/spl2/app/src/main/java/com/example/myapplication/CervicalActivity.java
@@ -59,7 +59,7 @@ public class CervicalActivity extends Fragment {
 
         myHelper = new MyHelper(getContext());
         sqLiteDatabase = myHelper.getWritableDatabase();
-        myHelper.deleteAll();
+
 
         lineDataSet1 = new LineDataSet(null,null);
         lineDataSet2 = new LineDataSet(null,null);
@@ -138,10 +138,15 @@ public class CervicalActivity extends Fragment {
 
                 int yVal = Integer.parseInt(value);
 
+                if(pointsAdded == false)
+                {
+                    myHelper.deleteAll();
+                }
+
                 myHelper.insertData(xVal, yVal);
 
                 lineDataSet.clear();
-                myHelper.deleteAll();
+
                 savedEntries = getData();
                 lineDataSet.setValues(getData());
                 lineDataSet.setLabel("Readings");
@@ -160,6 +165,8 @@ public class CervicalActivity extends Fragment {
                 graph.clear();
                 graph.setData(lineData);
                 graph.invalidate();
+
+                pointsAdded = true;
 
                 int x = Integer.parseInt(xInput.getText().toString());
                 x += 4;
@@ -223,5 +230,17 @@ public class CervicalActivity extends Fragment {
     public ArrayList<Entry> saveData(){
 
         return this.savedEntries;
+    }
+
+    public void refreshData(){
+        pointsAdded = false;
+        graph.clear();
+
+        lineData.removeDataSet(lineDataSet);
+        graph.setData(lineData);
+        graph.invalidate();
+        myHelper.deleteAll();
+        xInput.setText(Integer.toString(0));
+        yInput.setText("");
     }
 }
